Fix date_picker initial display value offset
On Firefox, the initial value set in the date_picker could be a day off
from the value stored in a user's database since browsers treat the
ISO 8601 date short form differently. This could be quite confusing for
users that see this bug, since the scenario would likely play out as
follows:

1. User sets date as 2017-07-01.
2. User saves and SQL updates to 2017-07-01.
3. User reopens form and sees 2017-06-30, the result of an improper date
   parse on the JS frontend.
4. Steps 1-3 are repeated until user realizes something is amiss and
   decides to submit a merge request.

In detail, Firefox sets the date in local time, and Chrome sets it as
UTC when only a date string is passed. The fix involves setting each
part of the date (year, month, date) explicitly.

The analysis above came from the help of a StackOverflow thread:
https://stackoverflow.com/questions/2587345/why-does-date-parse-give-incorrect-results