Set table field type value to null when all items are removed
This fixes a bug where you would never be able to remove the last
table row in a table field type. This was caused by the JavaScript
only updating the hidden value field whenever the number of rows
was > 0. This fix will update the hidden field value to null whenever
all rows are removed, as intended.