[BUG FIX] count() in PHP >= 7.2
In PHP >= 7.2 the count() argument needs to NOT be null. We must ensure we are running count against a valid countable argument, otherwise bail with 0.

We must check our codebase for instances of count, and sizeof (alias of count) and do the proper check for countable.

REF is_countable: [php manual](https://www.php.net/manual/en/function.is-countable.php)

This is backwards compatible.