made checklist field work inside repeatable field
Addresses #2266 . The checklist field now loads its options using PHP, but it checks those options using JavaScript. The value of the field is now stored in a hidden input, instead of individual checkboxes like it was doing before. The syncPivot() method has been amended to also support passing the connected entries as a JSON, not only a direct array. This was done because when storing multiple values inside one hidden input, the value that reaches PHP is a JSON. But previously, when there were individual inputs for each value, the value that reached PHP was a PHP array. This is a non-breaking change, and I expect it will also help with other fields - if we wish to store multiple IDs inside one hidden field.