select_and_order field - storing values as Select instead of hidden Input
In order to make this work inside the repeatable field, the values need to be stored inside one input. Multiple inputs won't do, because their names will be stripped out by the repeatable field. The only way to include multiple values inside one input is a Select multiple. So that's what I've done.