fixes #2797 - CrudColumn constructor did not guess missing attributes
This PR makes the makeSureColumnHasNeededAttributes() public instead of private, and makes sure that CrudColumns call that method inside the constructor, so that by the time it gets saved, it will have proper attributes. This makes fixes a problem (#2797) where you could not specify the label using the columns fluent syntax, because the key was different inside CrudColumn than it was inside the CrudPanel object. This also makes it easier to debug CrudColumns, because even after their first call, they will have all guessed attributes.