Fix lonely dot in CRUD when no entries
When there are zero entries in a DB table, the CRUD now shows an empty dot as the subtitle, so the actual heading reads "Categories . Reset" which is ugly and annoying. This PR fixes it my making the "infoEmpty" string on English "No entries". So now the heading shows "Categories. No entries. Reset". Much better I think - my OCD is happier now. While fixing this I've also changed a few lang entries for English - to call db records "entries" always, instead of sometimes "records" sometimes "entries". And I've added a fadeIn animation on the Reset button, so it matches the rest of the table, previously the button appeared to load slightly faster than the table.