also collect url and backpack_url as stats
Most of the time the new URL we send will give no extra information other than what we collected using SERVER variables - only the same information in a different format, that's easier to use as a human.  This will help us more easily identify the domain where Backpack is used, by looking at our stats. No GDPR implications since it doesn't have anything to do with the client, only the server. The backpack_url() is new, and allows us to double-check if the app still uses Backpack. In the past we've had cases where Backpack was removed from production but we couldn't tell that from our stats, we thought they're still using Backpack illegally. This should give us a quick way to double-check that they still are.