Datatables Persistent URL Tracking Fix
Under certain circumstances we want to render a table but not remember filtering. The flag for disabling table persistence is correct in that it doesn't trigger the redirect, however any changes made while that page is loaded still updates the localStorage value tracking filters so when a user returns to the regular persistence-enabled table it recreates the non-persistent state.

Update assures the the localStorage value code only runs when table persistence is enabled.